refactor(plugins): extract PackageManager from PluginManager (spec 053)
Split PluginManager into:
- PackageManager: generic distribution (download, install, update, registry, DB)
- PluginLoader: integration-specific (import, createPlugin, register)

Add type field to PluginManifest and plugins DB table.
Update registry.json with type: integration for all 10 plugins.
Renumber specs: 053-057 (was 053-056).
Add spec 054 for recipe packages.
