+++ specs/055-versioning-cicd-docker/spec.md
@@ -1,4 +1,4 @@
-# 054 — Versioning, CI/CD & Docker Images
+# 055 — Versioning, CI/CD & Docker Images
 
 ## Summary
