+++ specs/056-docker-influxdb-autoconfig/spec.md
@@ -1,4 +1,4 @@
-# 055 — Docker InfluxDB Auto-Configuration
+# 056 — Docker InfluxDB Auto-Configuration
 
 ## Summary
